:beers: api: moved model CertificationVersion to src/shared
