+++ api/src/certification/session/domain/models/Session.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { config } from '../../../../../lib/config.js';
-import { CertificationVersion } from '../../../../../lib/domain/models/CertificationVersion.js';
+import { CertificationVersion } from '../../../../shared/domain/models/CertificationVersion.js';
 
 const CREATED = 'created';
 const FINALIZED = 'finalized';
