+++ api/tests/integration/infrastructure/repositories/certification-challenge-repository_test.js
@@ -3,7 +3,7 @@ import { expect, knex, domainBuilder, databaseBuilder } from '../../../test-help
 import { CertificationChallenge } from '../../../../lib/domain/models/CertificationChallenge.js';
 import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
 import * as certificationChallengeRepository from '../../../../lib/infrastructure/repositories/certification-challenge-repository.js';
-import { CertificationVersion } from '../../../../lib/domain/models/CertificationVersion.js';
+import { CertificationVersion } from '../../../../src/shared/domain/models/CertificationVersion.js';
 
 describe('Integration | Repository | Certification Challenge', function () {
   describe('#save', function () {
