+++ api/tests/unit/application/certification-centers/certification-center-controller_test.js
@@ -1,11 +1,8 @@
-import { domainBuilder, expect, hFake, sinon } from "../../../test-helper.js";
-import { usecases } from "../../../../lib/domain/usecases/index.js";
-import {
-  certificationCenterController
-} from "../../../../lib/application/certification-centers/certification-center-controller.js";
+import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
+import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { certificationCenterController } from '../../../../lib/application/certification-centers/certification-center-controller.js';
 
 describe('Unit | Controller | certifications-center-controller', function () {
-
   describe('#getStudents', function () {
     it('should return a paginated serialized list of students', async function () {
       // given
