+++ api/tests/unit/domain/usecases/get-next-challenge-for-certification_test.js
@@ -1,7 +1,7 @@
 import { domainBuilder, expect, sinon, catchErr } from '../../../test-helper.js';
 import { getNextChallengeForCertification } from '../../../../lib/domain/usecases/get-next-challenge-for-certification.js';
 import { Assessment } from '../../../../lib/domain/models/Assessment.js';
-import { CertificationVersion } from '../../../../lib/domain/models/CertificationVersion.js';
+import { CertificationVersion } from '../../../../src/shared/domain/models/CertificationVersion.js';
 import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
 
 describe('Unit | Domain | Use Cases | get-next-challenge-for-certification', function () {
